Python/sanction_1/requirements.txt
streamlit
python-dotenv
pyodbc
pandas
rapidfuzz
phonetics
openai
